refactor(frontend): shorten ui logic in SideTab
