+++ frontend/src/components/sidetab/SideTabButton.tsx
@@ -4,15 +4,17 @@ import { SquareLineMenuIcon } from "@/assets/icons/SquareLineMenuIcon";
 import { NewSearchTabMenuIcon } from "@/assets/icons/NewSearchTabMenuIcon";
 import { UserMenuIcon } from "@/assets/icons/UserMenuIcon";
 
+type IconType = "article" | "history" | "newTab" | "settings";
+
 type props = {
-	buttonType: string | null;
-	buttonName: string | null;
+	name: string;
+	iconType?: IconType;
 };
 
 export function SideTabButton(props: props) {
-	const { buttonType, buttonName } = props;
-	const changeButtonIcons = (buttonType: string | null) => {
-		switch (buttonType) {
+	const { iconType, name } = props;
+	const changeIcons = (iconType: IconType | undefined) => {
+		switch (iconType) {
 			case "article":
 				return <BookmarksMenuIcon />;
 			case "history":
@@ -25,23 +27,6 @@ export function SideTabButton(props: props) {
 				return <SquareLineMenuIcon />;
 		}
 	};
-	const changeButtonName = (
-		buttonType: string | null,
-		buttonName: string | null,
-	) => {
-		switch (buttonType) {
-			case "article":
-				return "記事ブックマーク";
-			case "history":
-				return "検索履歴ブックマーク";
-			case "newTab":
-				return "新しい検索";
-			case "settings":
-				return "mail@mail.com";
-			default:
-				return buttonName;
-		}
-	};
 
 	return (
 		// TODO: change next/link component
@@ -49,10 +34,8 @@ export function SideTabButton(props: props) {
 			type="button"
 			className="px-8 py-4 w-full flex items-center justify-start gap-6 rounded-sm hover:bg-lrz-border-gray"
 		>
-			{changeButtonIcons(buttonType)}
-			<h1 className="font-zen font-bold tracking-wider">
-				{changeButtonName(buttonType, buttonName)}
-			</h1>
+			{changeIcons(iconType)}
+			<span className="font-zen font-bold tracking-wider">{name}</span>
 		</button>
 	);
 }
